+++ dessia_common/utils/helpers.py
@@ -10,8 +10,7 @@ That way, we can avoid cyclic imports.
 
 import sys
 from importlib import import_module
-from typing import Generic, TypeVar
-import inspect
+from typing import TypeVar
 
 _PYTHON_CLASS_CACHE = {}